::: POTENTIALLY BREAKING/CONFUSING CHANGE ::: TL;DR; any use of "enabled" -> "visible" (unless on a Batcher, batcher.enabled is correct, geometry.enabled is not)
Any use of "enabled" on geometry or text etc anywhere has been changed to "visible". The context in which it was used was visibility, Batcher uses it correctly where processing is ignored on disabled batchers. Any visibility related flags and values have been made consistent across the board. Sorry if this causes some minor code changes but it's unavoidable pre 1.0 :> Search your code for "enabled" and if you see any that aren't your own or on a batcher it's likely meant to be .visible instead.

General; part of #77, cleanup enabled -> visible
Matrix4; decompose was using an Array<Dynamic> as a return, now typed and proper, added "getPosition" as a convenience but this is inconsistent a bit still.. still deciding on this file.
BBST; left/right default to null for dynamic platforms that don't populate this for clarity/debugging
DebugView; profiler now has "offsets" that allow you to ignore profiles in other profiles. For example, the core.render loop includes the very debug view showing you that, which, when expensive, is masking the game rendering profile quite badly causing it to seem way worse. This basically says core.render is core.render - debugview. Profiler view also now has hidden views, that will track but not draw (like when tracking portions of the render stage that aren't relevant to the game, like debug views).