Scene; don't init entities from the defer list if they're inited already
The situation can arise from children added to parents after scene init
but before the parent init, meaning the parent will init the child, and
it lands in the defer list. There are also other places to have this
happen so the fix isn't a bandaid it's specifically ensuring correct
behaviour. logging was added for the duplicate case for debugging
potential errors masked by this (but multiple inits are much worse of
an error anyway)