Haxe 3.3; Workaround for 3.3.0-rc1 bug with inline constructors
Haxe 3.3.0-rc1 has a bug - when inline constructors are used on values
returned from inline functions and not stored in a local variables,
they will be typed with their initializer and not the actual
declaration type.

If class has field declared as `var x : Float = 0;` temporary variable
will have type Int as it's type of variable initializer.

(haxe issue pending)